WIP: Support for rectangular blits.
This works 99.9% of the time.  Where it seems to fail at present is when
blitting a byte-aligned rectangle to another byte-aligned rectangle in
the destination.  In this case, we get zeros for all bytes on the
right-hand edge of the rectangle.

This seems to be due to the interaction between two things:
1. When the right-hand edge aligns with a byte boundary, the
   first-column mask is set to zero, and,
2. When performing a blit, the first byte fetched will be ANDed with the
   first-column mask, yielding ... zero.

This is especially evident when blitting rectangles on byte boundaries
that are only 8 pixels wide.  In fact, that's exactly the unit test case
which happens to be failing at the moment.