introduce the Draw trait.
Before this patch, the various drawing operations on a stencil are
tightly bound to the Stencil type.  After this patch, any Stencil-like
thing could implement a Draw trait to facilitate different kinds of
stencils.

The motivation for this patch was an ill-fated "AutoStencil" type which
was intended to update the SDL frame buffer after every Draw operation.
It turns out that this would not type check (for some reason, and
frankly, I'm just too lazy to care at as of this commit).  So, I got rid
of the AutoStencil and I'm back to just normal stencils.  However, I
kept the trait in anticipation that I'll revisit this design choice
later.