Use per-gadget event bits to trigger actions.
Until now, gadgets had no way to communicate actions to the containing
application.  This patch adds an easy to use mechanism for an app to
react to a user input gesture.

Note that I use a 16-bit mask of bits to indicate events.  This allows
each gadget to have up to 16 unique kinds of gadget-specific events that
it can generate.  How the application reacts to these events will depend
on the exact type of the gadget.  For events which need more information
bandwidth than "something happened", an event bit can be used to trigger
the app to read from a gadget-specific interface to acquire more
details.

I don't use enumerations because they are (1) type-specific and (2) Not
expandable without altering the original source code.  Thus, I'm forced
to use the most type-generic representation I know: a simple bit-mask.

NOTE: Right now, paint requests are handled in a similar-but-different
manner.  I'm tempted to unify paint requests with this new, more general
purpose, event delivery system.