Implement simple toggle buttons using push buttons
OK, by sheer accident, it seems that a *bug* in the push button code
works in my favor to implement toggle buttons with an acceptable level
of code complexity.  :)

This does mean, however, that a proper push button needs to be inverted
in its click handler; otherwise it will remain "selected".  (This can
also be used to an advantage to indicate a lengthy operation as well.)