Refactor files.
I'm still not happy with calling the nucleus ... nucleus.

The nucleus *includes* the HAL; it's not a separate thing.  But, I also
don't want to call it "kernel" either, because nucleus and kernel are
(for our purposes) synonyms; therefore, a kernel *includes* a HAL by
definition.

What else can I call the damn thing?