Port DDCG compiler experiment to 65816
This is just a test to see how easy it is to port the DDCG concept to the
65816.

It's easy.  I was able to retarget the whole code generator back-end (such as
it is) in about a day.  Code produced is not optimal; but, it looks like it
should work correctly (except for the case where zero-page is used, where in
some cases it can become corrupted because it's not preserved before use.  This
is a known issue, but as of right now, too lazy to fix).