WIP: Introduce basic symbol table design.
This symbol table is sufficient to track a single procedure's local
symbols and function arguments.  The type of all symbols is assumed to
be 16-bits wide.  Direct page location 0 is assumed to be reserved for
use as a result value.

As of this patch, the symbol table is not yet used by the compiler
logic.