Recognize and react to the DC mnemonic
Prior to this patch, the assembler's input was just the operand for the
DC mnemonic.

After this patch, a complete assembly language line of input is
recognized, broken into its sub-components, and from there, we assemble
the DC operand.

This was done to better support the negative tests that still need to be
made to pass.