Restructure `HandleTable` <-> `EmState` relation
WARNING: This code still doesn't compile - see further
details below in the last paragraph.

Previously, `HandleTable` and `EmState` had a circular dependency
on each other, which has caused borrowing issues.

The new relationship is as follows:
- each `EmState` has a unique process id
- this process id is associated with it's own `HandleTable`
- all of these associations are stored in a new `HandleTableRepo`

                HandleTableRepo

EmStateUniqueProcessId  ->  HandleTable,
EmStateUniqueProcessId  ->  HandleTable,
EmStateUniqueProcessId  ->  HandleTable,
            ...         ...     ...

Using the unique process id (which implements `Copy`) as a key to
a `HandleTable` has the advantage that we don't need to store the
whole `EmState` for the association, which prevents borrowing issues.

Note: For generating the above mentioned process ids, we rely on
the (well established) `snowflake` crate:
https://crates.io/crates/snowflake

WARNING: There is still one piece of code which doesn't compile,
which has to do with crossing the FFI boundary to `sdl2`:
When trying to add a timer to `TimerSubsystem`, it's closure is
required to be `Send`, but our passed in closure doesn't fulfill
this requirement, because it must use `*mut ()` pointers due to
another `sdl2`-API and `*mut ()` pointers are not `Send`:
```rust
timer_subsystem.add_timer(
    period,
    Box::new(move || {
        //   ^^^ this closure needs to be `Send`
        let _ = event_subsystem
            .push_event(Event::User {
                timestamp: 0,
                window_id: 0,
                type_: timer_tick,
                code: signal.bit() as i32,
                data1: 0 as *mut libc::c_void,
                        //  ^^^^^^^^^^^^^^^^^^ this is not `Send`
                data2: 0 as *mut libc::c_void,
                        //  ^^^^^^^^^^^^^^^^^^ this is not `Send`
            })
            .unwrap();

        period
    }),
)
```
This seems to be a __fundamental__ incompatibility between the APIs
of `sdl2`. To resolve this, a new API __within `sdl2`__ needs to be
provided that turns the following API...
```rust
enum Event {
    // ...
    // some other omitted variants
    // ...
    User {
        timestamp: u32,
        window_id: u32,
        type_: u32,
        code: i32,
        data1: *mut c_void,
        data2: *mut c_void,
    },
}
```
...into roughly this new API...:
```rust
enum Event {
    // ...
    // some other omitted variants
    // ...
    User {
        timestamp: u32,
        window_id: u32,
        type_: u32,
        code: i32,
        data: Data
    },
}
```
...where `Data` is a wrapper struct that is provided by `sdl2`
and implements `Send`.